WIP: Reworked UI to be cleaner. Separation of MainActivity will commence soon(TM)
