--- app/app/src/main/java/com/python/pythonator/ui/main/view/SectionsPagerAdapter.java
@@ -1,37 +1,0 @@
-package com.python.pythonator.ui.main.view;
-
-import androidx.annotation.Nullable;
-import androidx.fragment.app.Fragment;
-import androidx.fragment.app.FragmentManager;
-import androidx.fragment.app.FragmentPagerAdapter;
-
-import com.python.pythonator.ui.queue.QueueFragment;
-
-public class SectionsPagerAdapter extends FragmentPagerAdapter {
-
-    private static final String[] TAB_TITLES = new String[]{"Queue"};
-
-    public SectionsPagerAdapter(FragmentManager fm) {
-        super(fm);
-    }
-
-    @Override
-    public Fragment getItem(int position) {
-        switch (position) {
-            case 0:
-            default:
-                return new QueueFragment();
-        }
-    }
-
-    @Nullable
-    @Override
-    public CharSequence getPageTitle(int position) {
-        return TAB_TITLES[position];
-    }
-
-    @Override
-    public int getCount() {
-        return 1;
-    }
-}
