+++ app/app/src/main/res/menu/main_menu.xml
@@ -5,12 +5,15 @@
         android:id="@+id/main_menu_bluetooth"
         android:icon="@drawable/ic_bluetooth_disabled"
         android:title="barcode"
-        app:showAsAction="always">
-    </item>
+        app:showAsAction="always"/>
+
+    <item android:id="@+id/main_menu_sort"
+        android:icon="@drawable/ic_sort"
+        android:title="sort"
+        app:showAsAction="always"/>
 
     <item android:id="@+id/main_menu_settings"
         android:title="settings"
         android:icon="@drawable/ic_settings"
-        app:showAsAction="always">
-    </item>
+        app:showAsAction="always"/>
 </menu>
